Switch to more generic testing structure
Previously, we had host tests and "unit tests".
The unit tests were ports to be triggered by test-runner, and the host
tests were programs to be invoked on the host.

This patch reshapes that so that tests are a test name and a command to run.
We also add convenience constructors (hosttest, boottest) to reproduce
previous functionality.

Change-Id: Ic130828a7c90875d6fc22a089b21d25fb651c286
Bug: 122330505