Run storage tests from Android userspace
With an adapted storageproxyd on the Android side, and RPMB over VirtIO
serial on the emulator side, it is now possible to run the "td" and
other storage-using tests on the emulator.

The td.init/td.check/td.clean tests need to wait until the emulator
supports keeping a userdata image from run to run.

We likely should have been running tdea tests all along, this enables
them since we are enabling other storage tests anyways.

Bug: 124277696
Change-Id: I144e9c0de00a9c16d9aab16b9829acad9df48ea3