Split test-map out of build-config
By splitting the test map content out of build-config, we support
alternate build-configs which contain a subset of targets - those
configurations can include the test-map without needing to register all
the targets.

This is helpful in the case that AOSP gets a new target that another
configuration or tree will not support, as it keeps build.py able to
build "all" targets.

Bug: 141883290
Change-Id: I794472079aed44b76065bb1f8f0a670ba033a884