scripts: Make envsetup.sh work from arbitrary dir
We want to be able to source envsetup.sh from locations other than the
trusty top dir so that this script can be used to find hermetic python3
binaries from python scripts. The gettop function borrows from a bash
function with the same name in Android's envsetup.sh. Also adds PY3
variable which points to the prebuilt hermetic python3 binary.

Bug: 197031257
Change-Id: I5e10c31827f42677d27da5f0594c9aec1209d00a