Roll clang forward to r445002 to match AOSP
The AOSP Rust 1.59 toolchain is built using clang r445002, and we need
to match this version so that the Rust host libraries can build
correctly. The Rust standard library crates in AOSP are built with LTO
enabled, and the LLVM IR format is not stable across compiler revisions.
Attempting to build against the Rust host libraries using a different
compiler revision is unlikely to work, so we need to stay in sync.

Test: build.py
Bug: 229330061
Change-Id: I8b205a479b29129288e2e5025e3ec94041f2a02d