scripts: Detect flaky tests by retrying them
Much time is lost to flaky tests. This change retries the first 10
tests that fail and flags them as potentially flaky if they pass
on the second try. This is not intended as a substitute for fixing
the root cause of flakyness; merely as way to detect problems and
avoid having to restart long presubmit runs due to spurious test
failures.

Bug: 236898842
Change-Id: I84b54060395c8dfcf3626f39ee1f32436b17c5b2