scripts: Prepare tests to run in the same QEMU instance
A few Android tests set the ro.boot.verifiedbootstate property to
fake. However, if the property is already set, the second attempt
to set it fails which means the whole test is treated as a failure.

Test whether the property was already set before trying to set it
such that we do not need to reboot before tests that set the
property.

Bug: 242077104
Change-Id: I41a27828544c0fc2b30a4a6b4cdb0a3ac9a3e495