scripts: Only retry tests in batch mode
It isn't all that helpful to retry tests during interactive
development but quite helpful when running tests in batch mode.

This CL defines batch mode as runs where there is no test filter
and where the debug on error flag is not set. If those conditions
are not set, the test is likely being run directly by the developer
and we should not retry it on failure to shorten the run time and
reduce the output from the script.

Bug: 236898842
Test: manually, run build.py qemu-generic-arm64-test-debug \
	--test $TESTNAME where the named test is modified to fail.
Change-Id: If7b58b35302c21ae2dac934691a9864b882eb784